Add another blackbox test.
Note: this test is red, which is really confusing.
I know for sure that my production code does not cover this scenario correctly.
But also I was convinced that the test implementation is too simple and would not catch this.
The behaviour of the new test is unexpected and needs more investigation.