Simplified the makefile for the blackbox tests for Cat.
Note: The previous version was correct, although only by accident.
I had tried hard to find a scenario which would proove that the test implementation is wrong, but did not succeed.
Due to my misunderstanding of the bash syntax, I had implemented the rule differently than I had meant;
but this healed the wrong rule...
Very confusing, all this.

Note: the Cat test cases are still green, except the new test case which is still red.
Everything according to the expectation now."