Correct the unit test implementation (green again).
Note: the expected value for noLines had an additional trailing line separator.
This bug was introduced with a previous refactoring, but remained undetected,
because also the input values of this test case were broken with the same pattern,
so effectively there was no unit test with noLines for a while.
Luckily, the production code was still working correctly.