Resolve it - green now.
This code is really ugly:
(1) The lambda does not accept a checked expression, so I must either suppress it, or wrap into an unchecked expression.
(2) I could not manage to skip one entry with exception but then continue the stream loop.
(3) I need to capture the error status; local variable does not work, it has to be an object field
    (and, yes, it should not be static, at least this I can improve)
(4) When I suppress the exception, the method must return something, but I must suppress this entry from the final output.