Roll back to commit ID 46f241b11a406357629871b24d659294470e09d6.
Note: I had tried to separate building of the headlines from reading the file contents.
I used separate streams for this.
However, this was not a good idea, because I need too much overhead to join these two streams later.
Also the structure of two streams is quite complex in comparison with the simple sequential pipeline we had before.