Roll back to commit ID fa5e22b.
Notes: (1) the introduction of the class hierarchy was done too early;
Cat and Head should first be unified much more.
The class hierarchy introduces another dimension of coupling,
which makes further refactorings too difficult.
(2) special attention of the data types of method parameters and fields.
The strategies I had tired to unify these data types lead to nowhere.
A better strategy is needed for this (maybe a dependency inversion / inversion of control)?